test(server): verify general tests without embedded package data
Run general tests and benchmarks in CI with an empty `pkgdata.zip`
overlay to catch dependencies on bundled package exports or documentation.
Mark tests that require real spx data with a test-only build constraint,
and check that the isolation overlay is active.

Centralize spx fixtures and keep package-independent type tests and the
alternative framework's `.spx` registration regression in the isolated
suite. Continue collecting coverage from the full test suite.

Add large-list document-link tests for plain XGo and registered classfiles,
plus an isolated benchmark. Retain the spx resource-scanning benchmark and
assert resource links before timing it.
