+++ internal/server/spx_definition_test.go
@@ -1,3 +1,5 @@
+//go:build !test_no_pkgdata
+
 package server
 
 import (
