+++ internal/server/spx_util_test.go
@@ -10,124 +10,6 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestIsInSpxPkg(t *testing.T) {
-	t.Run("NilObject", func(t *testing.T) {
-		assert.False(t, IsInSpxPkg(nil))
-	})
-
-	for _, tt := range []struct {
-		name string
-		pkg  *gotypes.Package
-		want bool
-	}{
-		{name: "NoPackage"},
-		{name: "OtherPackageNamedSpx", pkg: gotypes.NewPackage("example.com/spx", "spx")},
-		{name: "SpxSubpackage", pkg: gotypes.NewPackage(SpxPkgPath+"/subpkg", "subpkg")},
-		{name: "SpxPackage", pkg: GetSpxPkg(), want: true},
-		{name: "DistinctSpxPackage", pkg: gotypes.NewPackage(SpxPkgPath, "spx")},
-	} {
-		t.Run(tt.name, func(t *testing.T) {
-			obj := gotypes.NewTypeName(token.NoPos, tt.pkg, "Value", gotypes.Typ[gotypes.Int])
-			assert.Equal(t, tt.want, IsInSpxPkg(obj))
-		})
-	}
-}
-
-func TestGetSimplifiedTypeString(t *testing.T) {
-	for _, tt := range []struct {
-		name string
-		pkg  *gotypes.Package
-		want string
-	}{
-		{name: "NoPackage", want: "Value"},
-		{name: "OtherPackage", pkg: gotypes.NewPackage("example.com/sample", "sample"), want: "sample.Value"},
-		{name: "OtherPackageNamedSpx", pkg: gotypes.NewPackage("example.com/spx", "spx"), want: "spx.Value"},
-		{name: "SpxPackage", pkg: GetSpxPkg(), want: "Value"},
-		{name: "DistinctSpxPackage", pkg: gotypes.NewPackage(SpxPkgPath, "spx"), want: "spx.Value"},
-	} {
-		t.Run(tt.name, func(t *testing.T) {
-			named := gotypes.NewNamed(gotypes.NewTypeName(token.NoPos, tt.pkg, "Value", nil), gotypes.NewStruct(nil, nil), nil)
-			assert.Equal(t, tt.want, GetSimplifiedTypeString(named))
-		})
-	}
-}
-
-func TestResolvedNamedType(t *testing.T) {
-	pkg := gotypes.NewPackage("example.com/pkg", "pkg")
-	named := gotypes.NewNamed(gotypes.NewTypeName(token.NoPos, pkg, "Point", nil), gotypes.NewStruct(nil, nil), nil)
-	aliasToNamed := gotypes.NewAlias(gotypes.NewTypeName(token.NoPos, pkg, "PointAlias", nil), named)
-	aliasChainToNamed := gotypes.NewAlias(gotypes.NewTypeName(token.NoPos, pkg, "PointAliasChain", nil), aliasToNamed)
-	aliasToBasic := gotypes.NewAlias(gotypes.NewTypeName(token.NoPos, pkg, "StringAlias", nil), gotypes.Typ[gotypes.String])
-	aliasToPointerNamed := gotypes.NewAlias(gotypes.NewTypeName(token.NoPos, pkg, "PointPtrAlias", nil), gotypes.NewPointer(named))
-	aliasChainToPointerNamed := gotypes.NewAlias(gotypes.NewTypeName(token.NoPos, pkg, "PointPtrAliasChain", nil), aliasToPointerNamed)
-
-	for _, tt := range []struct {
-		name string
-		typ  gotypes.Type
-		want *gotypes.Named
-	}{
-		{
-			name: "Nil",
-			typ:  nil,
-			want: nil,
-		},
-		{
-			name: "Named",
-			typ:  named,
-			want: named,
-		},
-		{
-			name: "PointerToNamed",
-			typ:  gotypes.NewPointer(named),
-			want: named,
-		},
-		{
-			name: "AliasToNamed",
-			typ:  aliasToNamed,
-			want: named,
-		},
-		{
-			name: "PointerToAliasToNamed",
-			typ:  gotypes.NewPointer(aliasToNamed),
-			want: named,
-		},
-		{
-			name: "AliasChainToNamed",
-			typ:  aliasChainToNamed,
-			want: named,
-		},
-		{
-			name: "Basic",
-			typ:  gotypes.Typ[gotypes.Int],
-			want: nil,
-		},
-		{
-			name: "AliasToBasic",
-			typ:  aliasToBasic,
-			want: nil,
-		},
-		{
-			name: "AliasToPointerNamed",
-			typ:  aliasToPointerNamed,
-			want: named,
-		},
-		{
-			name: "AliasChainToPointerNamed",
-			typ:  aliasChainToPointerNamed,
-			want: named,
-		},
-	} {
-		t.Run(tt.name, func(t *testing.T) {
-			got := resolvedNamedType(tt.typ)
-			if tt.want == nil {
-				assert.Nil(t, got)
-				return
-			}
-			assert.Same(t, tt.want, got)
-		})
-	}
-}
-
 func TestPropertyTargetNamedTypeForCall(t *testing.T) {
 	newNamed := func(pkg *gotypes.Package, name string) *gotypes.Named {
 		return gotypes.NewNamed(gotypes.NewTypeName(token.NoPos, pkg, name, nil), gotypes.NewStruct(nil, nil), nil)
@@ -283,6 +165,43 @@ func TestPropertyTargetNamedTypeForCall(t *testing.T) {
 			if assert.NotNil(t, got) {
 				assert.Equal(t, tt.wantName, got.Obj().Name())
 			}
+		})
+	}
+}
+
+func TestIsInSpxPkg(t *testing.T) {
+	t.Run("NilObject", func(t *testing.T) {
+		assert.False(t, IsInSpxPkg(nil))
+	})
+
+	for _, tt := range []struct {
+		name string
+		pkg  *gotypes.Package
+	}{
+		{name: "NoPackage"},
+		{name: "OtherPackageNamedSpx", pkg: gotypes.NewPackage("example.com/spx", "spx")},
+		{name: "SpxSubpackage", pkg: gotypes.NewPackage(SpxPkgPath+"/subpkg", "subpkg")},
+	} {
+		t.Run(tt.name, func(t *testing.T) {
+			obj := gotypes.NewTypeName(token.NoPos, tt.pkg, "Value", gotypes.Typ[gotypes.Int])
+			assert.False(t, IsInSpxPkg(obj))
+		})
+	}
+}
+
+func TestGetSimplifiedTypeString(t *testing.T) {
+	for _, tt := range []struct {
+		name string
+		pkg  *gotypes.Package
+		want string
+	}{
+		{name: "NoPackage", want: "Value"},
+		{name: "OtherPackage", pkg: gotypes.NewPackage("example.com/sample", "sample"), want: "sample.Value"},
+		{name: "OtherPackageNamedSpx", pkg: gotypes.NewPackage("example.com/spx", "spx"), want: "spx.Value"},
+	} {
+		t.Run(tt.name, func(t *testing.T) {
+			named := gotypes.NewNamed(gotypes.NewTypeName(token.NoPos, tt.pkg, "Value", nil), gotypes.NewStruct(nil, nil), nil)
+			assert.Equal(t, tt.want, GetSimplifiedTypeString(named))
 		})
 	}
 }
