+++ internal/server/util_test.go
@@ -1,6 +1,7 @@
 package server
 
 import (
+	gotypes "go/types"
 	"testing"
 
 	"github.com/goplus/xgo/ast"
@@ -713,6 +714,82 @@ func TestIsRangesOverlap(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			assert.Equal(t, tt.want, IsRangesOverlap(tt.a, tt.b))
 			assert.Equal(t, tt.want, IsRangesOverlap(tt.b, tt.a))
+		})
+	}
+}
+
+func TestResolvedNamedType(t *testing.T) {
+	pkg := gotypes.NewPackage("example.com/pkg", "pkg")
+	named := gotypes.NewNamed(gotypes.NewTypeName(token.NoPos, pkg, "Point", nil), gotypes.NewStruct(nil, nil), nil)
+	aliasToNamed := gotypes.NewAlias(gotypes.NewTypeName(token.NoPos, pkg, "PointAlias", nil), named)
+	aliasChainToNamed := gotypes.NewAlias(gotypes.NewTypeName(token.NoPos, pkg, "PointAliasChain", nil), aliasToNamed)
+	aliasToBasic := gotypes.NewAlias(gotypes.NewTypeName(token.NoPos, pkg, "StringAlias", nil), gotypes.Typ[gotypes.String])
+	aliasToPointerNamed := gotypes.NewAlias(gotypes.NewTypeName(token.NoPos, pkg, "PointPtrAlias", nil), gotypes.NewPointer(named))
+	aliasChainToPointerNamed := gotypes.NewAlias(gotypes.NewTypeName(token.NoPos, pkg, "PointPtrAliasChain", nil), aliasToPointerNamed)
+
+	for _, tt := range []struct {
+		name string
+		typ  gotypes.Type
+		want *gotypes.Named
+	}{
+		{
+			name: "Nil",
+			typ:  nil,
+			want: nil,
+		},
+		{
+			name: "Named",
+			typ:  named,
+			want: named,
+		},
+		{
+			name: "PointerToNamed",
+			typ:  gotypes.NewPointer(named),
+			want: named,
+		},
+		{
+			name: "AliasToNamed",
+			typ:  aliasToNamed,
+			want: named,
+		},
+		{
+			name: "PointerToAliasToNamed",
+			typ:  gotypes.NewPointer(aliasToNamed),
+			want: named,
+		},
+		{
+			name: "AliasChainToNamed",
+			typ:  aliasChainToNamed,
+			want: named,
+		},
+		{
+			name: "Basic",
+			typ:  gotypes.Typ[gotypes.Int],
+			want: nil,
+		},
+		{
+			name: "AliasToBasic",
+			typ:  aliasToBasic,
+			want: nil,
+		},
+		{
+			name: "AliasToPointerNamed",
+			typ:  aliasToPointerNamed,
+			want: named,
+		},
+		{
+			name: "AliasChainToPointerNamed",
+			typ:  aliasChainToPointerNamed,
+			want: named,
+		},
+	} {
+		t.Run(tt.name, func(t *testing.T) {
+			got := resolvedNamedType(tt.typ)
+			if tt.want == nil {
+				assert.Nil(t, got)
+				return
+			}
+			assert.Same(t, tt.want, got)
 		})
 	}
 }
